+++ prim_ctrl/__main__.py
@@ -728,7 +728,7 @@ class Funnel(Pingable):
         self.external_url = f'https://{machine_name}.{tailscale.tailnet}:{external_port}{local_path}'
         self.local_tailscale = local_tailscale
         self.external_public_dns_resolver = None
-        self.external_tailscale_dns_resolver = None
+        self.external_tailscale_dns_resolvers = None
 
     async def wait_for(self, available: bool, timeout: float):
         self._sleepcounter = 0
@@ -736,13 +736,17 @@ class Funnel(Pingable):
 
     async def ping(self, availability_hint: bool | None = None):
         logger.debug("Resolving DNS for %s (%s:%s)", LazyStr(self.get_class_name), self.external_name, self.external_port)
-        # first try at Tailscale's DNS, if it doesn't know, we should not resolve at a public DNS and cache nxdomain for 5 minutes
-        if self.external_tailscale_dns_resolver is None:
-            self.external_tailscale_dns_resolver = ExternalDnsResolver(await self.local_tailscale.external_tailscale_dns_resolver())
+        # first try at Tailscale's DNSs, if they don't know, we should not resolve at a public DNS and cache nxdomain for 5 minutes
+        if self.external_tailscale_dns_resolvers is None:
+            resolvers = [await self.local_tailscale.external_tailscale_dns_resolver()]
+            resolvers.extend(await self._external_tailscale_dns_resolvers())
+            self.external_tailscale_dns_resolvers = [ExternalDnsResolver(resolver) for resolver in resolvers]
         try:
-            _answer = await self.external_tailscale_dns_resolver.resolve(self.external_name, self.external_port)
+            _answer = await self.external_tailscale_dns_resolvers[0].resolve(self.external_name, self.external_port)
         except Exception as e:
-            logger.debug("Resolving at Tailscale's external DNS has failed: %s", LazyStr(exception_repr, e))
+            logger.debug("Resolving at Tailscale's external DNS %s has failed: %s", self.external_tailscale_dns_resolvers[0].where, LazyStr(exception_repr, e))
+            # move current resolver to the end of the list
+            self.external_tailscale_dns_resolvers.append(self.external_tailscale_dns_resolvers.pop(0))
             return False
         # then try at a public DNS
         if self.external_public_dns_resolver is None:
@@ -755,7 +759,7 @@ class Funnel(Pingable):
         return True
 
     async def _sleep_while_wait(self, available: bool):
-        if self.local_tailscale.is_started_now and 0 != self._sleepcounter and 0 == self._sleepcounter % 3:
+        if self.local_tailscale.is_started_now and 0 != self._sleepcounter and 0 == self._sleepcounter % 6:
             logger.info("Restarting %s to retrigger public DNS records' configuration at Tailscale...", LazyStr(self.local_tailscale.get_class_name))
             await self.local_tailscale.stop(10, 30)
             await self.local_tailscale.start(10, 30)
@@ -763,6 +767,10 @@ class Funnel(Pingable):
             logger.info("Waiting for public DNS records to be updated for %s (%s)...", LazyStr(self.get_class_name), self.external_name)
         await asyncio.sleep(10)
         self._sleepcounter += 1
+
+    async def _external_tailscale_dns_resolvers(self):
+        logger.debug("Getting external Tailscale DNS resolvers for ts.net")
+        return [str(rr.to_text()) for rr in await dns.asyncresolver.Resolver().resolve("ts.net", rdtype=dns.rdatatype.NS)]
 
 class LocalTailscaleManager(Manager):
     async def start(self):