+++ CHANGELOG.md
@@ -3,6 +3,7 @@
 # vNext
 
 - Use cache for Tailscale devices API
+- Loop through all Tailscale DNSs when testing Funnel's DNS configuration
 
 ## 0.8.2